requirements.txt
# Core Dependencies
python-dotenv==1.0.0
requests==2.31.0

# Data Processing
pandas==2.1.3
numpy==1.26.2

# Web Scraping
beautifulsoup4==4.12.2
lxml==4.9.3
selenium==4.15.2

# Phone Number Validation
phonenumbers==8.13.26

# OCR (Optional - install if needed)
pytesseract==0.3.10
pdf2image==1.16.3
Pillow==10.1.0
easyocr==1.7.0

# NLP & Machine Learning (for enhanced enrichment)
transformers==4.35.2
torch==2.1.1
sentence-transformers==2.2.2
sentencepiece==0.1.99

# PDF Generation
reportlab==4.0.7
weasyprint==60.1

# Database
# SQLite is built-in
# For PostgreSQL migration (uncomment when needed):
# psycopg2-binary==2.9.9
# sqlalchemy==2.0.23

# Task Queue (for production scaling)
# celery==5.3.4
# redis==5.0.1

# API Framework (if building REST API)
# fastapi==0.104.1
# uvicorn==0.24.0
# pydantic==2.5.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0

# Code Quality
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Logging & Monitoring
colorlog==6.8.0